Changement des tables pour bug foreign key.

+++ database/migrations/2017_04_05_083007_create_partners_openings_table.php
@@ -15,7 +15,7 @@ class CreatePartnersOpeningsTable extends Migration
     {
         Schema::create('partners_openings', function(Blueprint $table) {
             $table->bigIncrements('id');
-            $table->integer('partner_id')->unsigned();
+            $table->bigInteger('partner_id')->unsigned();
             $table->foreign('partner_id')
                 ->references('id')
                 ->on('partners')
