+++ database/migrations/2017_04_20_110725_create_orders_info_temp_table.php
@@ -16,17 +16,17 @@ class CreateOrdersInfoTempTable extends Migration
         Schema::create('orders_info_temp', function (Blueprint $table){
             $table->bigIncrements('id');
             $table->timestamps();
-            $table->integer('applicationUser_id')->unsigned()->nullable();
+            $table->bigInteger('applicationUser_id')->unsigned()->nullable();
             $table->foreign('applicationUser_id')
                 ->references('id')
                 ->on('application_users')
                 ->onDelete('set null');
-            $table->integer('partner_id')->unsigned()->nullable();
+            $table->bigInteger('partner_id')->unsigned()->nullable();
             $table->foreign('partner_id')
                 ->references('id')
                 ->on('partners')
                 ->onDelete('set null');
-            $table->integer('order_id')->unsigned();
+            $table->bigInteger('order_id')->unsigned();
             $table->foreign('order_id')
                 ->references('id')
                 ->on('orders_info');
