+++ database/migrations/2017_05_20_212153_create_application_users_email_validation_table.php
@@ -16,7 +16,7 @@ class CreateApplicationUsersEmailValidationTable extends Migration
         Schema::create('application_users_email_validation', function (Blueprint $table) {
             $table->bigIncrements('id');
             $table->timestamps();
-            $table->integer('applicationUser_id')->unsigned();
+            $table->bigInteger('applicationUser_id')->unsigned();
             $table->foreign('applicationUser_id')
                 ->references('id')
                 ->on('application_users')
