+++ database/migrations/2017_10_03_174544_create_incident_table.php
@@ -16,7 +16,7 @@ class CreateIncidentTable extends Migration
         Schema::create('incidents', function(Blueprint $table) {
             $table->bigIncrements('id');
             $table->timestamps();
-            $table->integer('order_id')->unsigned()->nullable();
+            $table->bigInteger('order_id')->unsigned()->nullable();
             $table->foreign('order_id')
                 ->references('id')
                 ->on('orders_info')
