+++ database/migrations/2017_10_08_121331_create_incident_monitoring_table.php
@@ -16,7 +16,7 @@ class CreateIncidentMonitoringTable extends Migration
         Schema::create('incidents_monitoring', function(Blueprint $table) {
             $table->bigIncrements('id');
             $table->timestamps();
-            $table->integer('order_id')->unsigned()->nullable();
+            $table->bigInteger('order_id')->unsigned()->nullable();
             $table->foreign('order_id')
                 ->references('id')
                 ->on('orders_info')
